Splitting Net for windows and non windows compile
